backend/requirements.txt
fastapi
uvicorn
pillow
pytesseract
gtts
python-multipart
